refactor proof extraction logic into a `spl-token-confidential-transfer-proof-extraction`
