+++ token/confidential-transfer/proof-extraction/Cargo.toml
@@ -10,7 +10,9 @@ edition = "2021"
 [dependencies]
 bytemuck = "1.19.0"
 solana-curve25519 = "2.0.3"
+solana-program = "2.0.3"
 solana-zk-sdk = "2.0.3"
+spl-pod = "0.4.0"
 thiserror = "1.0.65"
 
 [lib]
